Refactor: reuse CustomOpenAIClient instead of cloning
- Add `extraBody` option to CustomOpenAIClient so subclasses can inject
  provider-specific request body params via the OpenAI SDK's extra body API
- Rewrite ZhipuOpenAIClient as a thin subclass (~15 lines vs ~250 cloned)
- Fix pre-existing bugs in CustomOpenAIClient:
  - `inputSchema` → `parameters` in tool definitions (OpenAI API spec)
  - Preserve original role for string-content system/assistant messages
